move typings into a separate directory
