--- app/index.d.ts
@@ -1,1 +1,0 @@
-// Dummy file, required by tsc
