+++ app/tsconfig.json
@@ -7,6 +7,8 @@
   },
   "include": [
     "./**/*",
-    "./package.json"
+    "./package.json",
+    "../typings/extend-electron.d.ts",
+    "../typings/ext-modules.d.ts"
   ]
 }
