+++ lib/actions/index.ts
@@ -1,6 +1,6 @@
 import rpc from '../rpc';
 import {INIT} from '../constants';
-import type {HyperDispatch} from '../hyper';
+import type {HyperDispatch} from '../../typings/hyper';
 
 export default function init() {
   return (dispatch: HyperDispatch) => {
