+++ lib/components/header.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {decorate, getTabsProps} from '../utils/plugins';
 
 import Tabs_ from './tabs';
-import type {HeaderProps} from '../hyper';
+import type {HeaderProps} from '../../typings/hyper';
 
 const Tabs = decorate(Tabs_, 'Tabs');
