+++ lib/components/notification.tsx
@@ -1,5 +1,5 @@
 import React, {forwardRef, useEffect, useRef, useState} from 'react';
-import type {NotificationProps} from '../hyper';
+import type {NotificationProps} from '../../typings/hyper';
 
 const Notification = forwardRef<HTMLDivElement, React.PropsWithChildren<NotificationProps>>((props, ref) => {
   const dismissTimer = useRef<NodeJS.Timeout | undefined>(undefined);
