+++ lib/components/style-sheet.tsx
@@ -1,5 +1,5 @@
 import React, {forwardRef} from 'react';
-import type {StyleSheetProps} from '../hyper';
+import type {StyleSheetProps} from '../../typings/hyper';
 
 const StyleSheet = forwardRef<HTMLStyleElement, StyleSheetProps>((props, ref) => {
   const {borderColor} = props;
