+++ lib/components/tab.tsx
@@ -1,5 +1,5 @@
 import React, {forwardRef} from 'react';
-import type {TabProps} from '../hyper';
+import type {TabProps} from '../../typings/hyper';
 
 const Tab = forwardRef<HTMLLIElement, TabProps>((props, ref) => {
   const handleClick = (event: React.MouseEvent) => {
