--- lib/ext-modules.d.ts
@@ -1,3 +1,0 @@
-declare module 'php-escape-shell' {
-  export function php_escapeshellcmd(path: string): string;
-}
