+++ lib/store/configure-store.prod.ts
@@ -5,7 +5,7 @@ import rootReducer from '../reducers/index';
 import effects from '../utils/effects';
 import * as plugins from '../utils/plugins';
 import writeMiddleware from './write-middleware';
-import type {HyperState, HyperActions} from '../hyper';
+import type {HyperState, HyperActions} from '../../typings/hyper';
 const thunk: ThunkMiddleware<HyperState, HyperActions> = _thunk;
 
 export default () =>
