+++ lib/utils/ipc.ts
@@ -1,4 +1,4 @@
 import {ipcRenderer as _ipc} from 'electron';
-import type {IpcRendererWithCommands} from '../../common';
+import type {IpcRendererWithCommands} from '../../typings/common';
 
 export const ipcRenderer = _ipc as IpcRendererWithCommands;
