+++ tsconfig.json
@@ -7,7 +7,8 @@
     "./app/",
     "./lib/",
     "./test/",
-    "./cli/"
+    "./cli/",
+    "./typings"
   ],
   "references": [
     {
