+++ typings/config.d.ts
